[feat] #27 create complete Promise API

+++ src/main/java/org/kkumulkkum/server/repository/PromiseRepository.java
@@ -3,5 +3,5 @@ package org.kkumulkkum.server.repository;
 import org.kkumulkkum.server.domain.Promise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromiseRepository extends JpaRepository<Promise, Integer> {
+public interface PromiseRepository extends JpaRepository<Promise, Long> {
 }
